importer/requirements.txt
weconnect[Images]~=0.60.3
PyYAML~=6.0.1
postgrest~=0.16.8
PyJWT~=2.8.0
sqlalchemy~=2.0.31
psycopg[binary,pool]~=3.2.1
haversine~=2.8.1
requests~=2.32.3
